nexus-backend (1)/nexus/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
langgraph==0.2.28
langchain-core==0.3.15
groq==0.9.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
python-docx==1.1.2
dateparser==1.2.0
python-dotenv==1.0.1
python-multipart==0.0.12
